Improve StateServiceStore
