+++ src/bctklib/persistence/ICacheSnapshot.cs
@@ -1,0 +1,18 @@
+// Copyright (C) 2015-2025 The Neo Project.
+//
+// ICacheSnapshot.cs file belongs to neo-express project and is free
+// software distributed under the MIT software license, see the
+// accompanying file LICENSE in the main directory of the
+// repository or https://opensource.org/license/MIT for more details.
+//
+// Redistribution and use in source and binary forms with or without
+// modifications are permitted.
+
+namespace Neo.BlockchainToolkit.Persistence;
+
+internal interface ICacheSnapshot : IDisposable
+{
+    void Add(ReadOnlyMemory<byte> key, byte[] value);
+    void Commit();
+}
+
